requirements.txt
Flask
python-dotenv
pdfplumber
pytesseract
Pillow
requests
